@@ -4,12 +4,21 @@ import * as vscode from "vscode";
 import * as functions from "./functions";
 
 export function activate(context: vscode.ExtensionContext) {
-  let openInputCommand = vscode.commands.registerCommand(
-    "pubspec-assist.openInput",
-    functions.openInput
-  );
+  let commands = {
+    addDependencyCommand: vscode.commands.registerCommand(
+      "pubspec-assist.addDependency",
+      (_: vscode.ExtensionContext) => functions.addDependency("dependencies")
+    ),
+    addDevDependencyCommand: vscode.commands.registerCommand(
+      "pubspec-assist.addDevDependency",
+      (_: vscode.ExtensionContext) =>
+        functions.addDependency("dev_dependencies")
+    ),
+  };
 
-  context.subscriptions.push(openInputCommand);
+  Object.values(commands).forEach((command) =>
+    context.subscriptions.push(command)
+  );
 }
 
 // this method is called when your extension is deactivated
